Chore: 초기 디렉토리 구조 변경 #5
- 도메인 디렉토리들을 domain으로 묶음
- 도메인 디렉토리의 하위 domain 디렉토리 이름은 entity로 변경
- common 디렉토리 추가
