--- src/main/java/com/ssginc/showpingrefactoring/cart/domain/CartId.java
@@ -1,20 +1,0 @@
-package com.ssginc.showpingrefactoring.cart.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Embeddable
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-public class CartId implements Serializable {
-
-    private Long memberNo;
-
-    private Long productNo;
-
-}
