+++ src/main/java/com/ssginc/showpingrefactoring/domain/cart/entity/CartId.java
@@ -1,0 +1,20 @@
+package com.ssginc.showpingrefactoring.domain.cart.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Embeddable
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+public class CartId implements Serializable {
+
+    private Long memberNo;
+
+    private Long productNo;
+
+}
