+++ src/main/java/com/ssginc/showpingrefactoring/domain/member/entity/MemberRole.java
@@ -1,0 +1,17 @@
+package com.ssginc.showpingrefactoring.domain.member.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum MemberRole {
+
+    ROLE_ADMIN("ROLE_ADMIN"),
+    ROLE_USER("ROLE_USER");
+
+    private final String role;
+
+    MemberRole(String role) {
+        this.role = role;
+    }
+
+}
