+++ src/main/java/com/ssginc/showpingrefactoring/domain/order/entity/OrderDetailId.java
@@ -1,0 +1,20 @@
+package com.ssginc.showpingrefactoring.domain.order.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Embeddable
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+public class OrderDetailId implements Serializable {
+
+    private Long productNo;
+
+    private Long ordersNo;
+
+}
