+++ src/main/java/com/ssginc/showpingrefactoring/domain/order/entity/OrderStatus.java
@@ -1,0 +1,18 @@
+package com.ssginc.showpingrefactoring.domain.order.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum OrderStatus {
+
+    READY("준비 중"),
+    TRANSIT("배송 중"),
+    COMPLETE("배송 완료");
+
+    private final String orderStatus;
+
+    OrderStatus(String orderStatus) {
+        this.orderStatus = orderStatus;
+    }
+
+}
