+++ src/main/java/com/ssginc/showpingrefactoring/domain/product/entity/Category.java
@@ -1,0 +1,25 @@
+package com.ssginc.showpingrefactoring.domain.product.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "category")
+public class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "category_no")
+    private Long categoryNo;
+
+    @Column(name = "category_name", length = 100)
+    private String categoryName;
+
+}
