+++ src/main/java/com/ssginc/showpingrefactoring/domain/report/entity/BlackListDelete.java
@@ -1,0 +1,17 @@
+package com.ssginc.showpingrefactoring.domain.report.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum BlackListDelete {
+
+    Y("Y"),
+    N("N");
+
+    private final String blackListDelete;
+
+    BlackListDelete(String blackListDelete) {
+        this.blackListDelete = blackListDelete;
+    }
+
+}
