+++ src/main/java/com/ssginc/showpingrefactoring/domain/report/entity/ReportStatus.java
@@ -1,0 +1,17 @@
+package com.ssginc.showpingrefactoring.domain.report.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum ReportStatus {
+
+    PROCEEDING("미처리"),
+    COMPLETED("처리");
+
+    private final String reportStatus;
+
+    ReportStatus(String reportStatus) {
+        this.reportStatus = reportStatus;
+    }
+
+}
