--- src/main/java/com/ssginc/showpingrefactoring/member/domain/MemberRole.java
@@ -1,17 +1,0 @@
-package com.ssginc.showpingrefactoring.member.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum MemberRole {
-
-    ROLE_ADMIN("ROLE_ADMIN"),
-    ROLE_USER("ROLE_USER");
-
-    private final String role;
-
-    MemberRole(String role) {
-        this.role = role;
-    }
-
-}
