--- src/main/java/com/ssginc/showpingrefactoring/member/domain/Redis.java
@@ -1,18 +1,0 @@
-package com.ssginc.showpingrefactoring.member.domain;
-
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.redis.core.RedisHash;
-
-@RedisHash(value = "MemberToken", timeToLive = 3600 * 24 * 14)
-@AllArgsConstructor
-@Getter
-@Setter
-public class Redis {
-
-    @Id
-    private String email;
-    private String refreshToken;
-}
