--- src/main/java/com/ssginc/showpingrefactoring/member/domain/SocialType.java
@@ -1,15 +1,0 @@
-package com.ssginc.showpingrefactoring.member.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum SocialType {
-
-    NAVER("네이버"),
-    KAKAO("카카오");
-
-    private final String socialType;
-
-    SocialType(String socialType) { this.socialType = socialType; }
-
-}
