--- src/main/java/com/ssginc/showpingrefactoring/order/domain/OrderDetailId.java
@@ -1,20 +1,0 @@
-package com.ssginc.showpingrefactoring.order.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Embeddable
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-public class OrderDetailId implements Serializable {
-
-    private Long productNo;
-
-    private Long ordersNo;
-
-}
