--- src/main/java/com/ssginc/showpingrefactoring/payment/domain/PaymentStatus.java
@@ -1,19 +1,0 @@
-package com.ssginc.showpingrefactoring.payment.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum PaymentStatus {
-
-    READY("대기"),
-    COMPLETE("완료"),
-    FAILURE("실패"),
-    CANCEL("취소");
-
-    private final String paymentStatus;
-
-    PaymentStatus(String paymentStatus) {
-        this.paymentStatus = paymentStatus;
-    }
-
-}
