--- src/main/java/com/ssginc/showpingrefactoring/product/domain/Category.java
@@ -1,25 +1,0 @@
-package com.ssginc.showpingrefactoring.product.domain;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "category")
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "category_no")
-    private Long categoryNo;
-
-    @Column(name = "category_name", length = 100)
-    private String categoryName;
-
-}
