--- src/main/java/com/ssginc/showpingrefactoring/report/domain/ReportStatus.java
@@ -1,17 +1,0 @@
-package com.ssginc.showpingrefactoring.report.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum ReportStatus {
-
-    PROCEEDING("미처리"),
-    COMPLETED("처리");
-
-    private final String reportStatus;
-
-    ReportStatus(String reportStatus) {
-        this.reportStatus = reportStatus;
-    }
-
-}
