--- src/main/java/com/ssginc/showpingrefactoring/report/domain/ReportType.java
@@ -1,17 +1,0 @@
-package com.ssginc.showpingrefactoring.report.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum ReportType {
-
-    CHAT("채팅"),
-    REVIEW("리뷰");
-
-    private final String reportType;
-
-    ReportType(String reportType) {
-        this.reportType = reportType;
-    }
-
-}
