--- src/main/java/com/ssginc/showpingrefactoring/stream/domain/StreamStatus.java
@@ -1,16 +1,0 @@
-package com.ssginc.showpingrefactoring.stream.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum StreamStatus {
-    STANDBY("송출 대기"),
-    ONAIR("송출 중"),
-    ENDED("송출 완료");
-
-    private final String streamStatus;
-
-    StreamStatus(String streamStatus) {
-        this.streamStatus = streamStatus;
-    }
-}
